Keep Qwen exporters recipe-local
